+++ Backend/src/main/resources/db/migration/V5__Create_Kategorite_Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Kategorite (
-    kategori_id INT IDENTITY(1,1) PRIMARY KEY,
+    kategori_id BIGINT IDENTITY(1,1) PRIMARY KEY,
     emri_kategorise NVARCHAR(100) NOT NULL,
     pershkrimi NVARCHAR(200),
     aktive BIT DEFAULT 1
